try to introduce some constants and get rid of magic numbers in code
50, 70 ,80 these kind of magic numbers make code hard to understand,
especially for people who did not write them. Even the programmer
himself might forget what they mean after 1-2 months, So it is better to
give meaning to numbers and use constants whenever a number is repeated
throughout code. This gives meaning to the number and makes code easier
to comprehend.